+++ src/unit-tests/test.py
@@ -66,7 +66,7 @@ class Test_PyOOCalc_Document(unittest.TestCase):
 
     def test_document_initialize(self):
         doc = pyoocalc.Document()
-        self.assertFalse(doc.is_null())
+        self.assertFalse(doc.is_null)
 
     def test_document_new_save_close(self):
         doc = pyoocalc.Document()
@@ -87,11 +87,11 @@ class Test_PyOOCalc_Document(unittest.TestCase):
 
     @pyoocalc_open_close_doc
     def test_document_sheets(self, doc):
-        self.assertFalse(doc.sheets.is_null(), "get sheets object")
+        self.assertFalse(doc.sheets.is_null, "get sheets object")
 
     @pyoocalc_open_close_doc
     def test_document_fields(self, doc):
-        self.assertFalse(doc.fields.is_null(), "get fields object")
+        self.assertFalse(doc.fields.is_null, "get fields object")
 
 ###############################################################################
 
@@ -119,10 +119,10 @@ class Test_PyOOCalc_Fields(Test_PyOOCalc_Base):
 
     def test_fields_field(self):
         field = self._doc.fields.field("TABLE_NAME")
-        self.assertFalse(field.is_null(), "get field object")
+        self.assertFalse(field.is_null, "get field object")
 
     def test_fields_count(self):
-        self.assertEqual(self._doc.fields.count(), 11,
+        self.assertEqual(self._doc.fields.count, 11,
                          "Wrong number of fields")
 
 ###############################################################################
@@ -174,14 +174,14 @@ class Test_PyOOCalc_Sheets(Test_PyOOCalc_Base):
 
     def test_sheets_sheet_by_index(self):
         sheet = self._doc.sheets.sheet("Sheet1")
-        self.assertFalse(sheet.is_null(), "get sheet object")
+        self.assertFalse(sheet.is_null, "get sheet object")
 
     def test_sheets_insert_remove_spreadsheet_count(self):
         self.assertTrue(self._doc.sheets.insert_spreadsheet("test1", 1))
-        self.assertEqual(self._doc.sheets.count(), 2,
+        self.assertEqual(self._doc.sheets.count, 2,
                          "Wrong number of fields")
         self.assertTrue(self._doc.sheets.remove_spreadsheet("test1"))
-        self.assertEqual(self._doc.sheets.count(), 1,
+        self.assertEqual(self._doc.sheets.count, 1,
                          "Wrong number of fields")
 
 ###############################################################################